+++ views/layouts/footer.php
@@ -46,6 +46,12 @@
     <script src="<?= BASE_URL ?>/templates/AdminLTE-3.2.0/plugins/jquery-validation/jquery.validate.min.js"></script>
     <script src="<?= BASE_URL ?>/templates/AdminLTE-3.2.0/plugins/jquery-validation/localization/messages_es.min.js"></script>
 <?php endif; ?>
+<!-- Page specific scripts -->
+<?php if (isset($pageScripts)): ?>
+    <?php foreach ($pageScripts as $script): ?>
+        <script src="<?= BASE_URL; ?><?= $script ?>"></script>
+    <?php endforeach; ?>
+<?php endif; ?>
 <!-- UI Components Utils (General UI Helpers) -->
 <script src="<?= BASE_URL; ?>/js/core/ui-components.js"></script>